\name{NewPackage-package}
\alias{NewPackage-package}
\alias{NewPackage}
\docType{package}
\title{
\packageTitle{NewPackage}
}
\description{
\packageDescription{NewPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NewPackage}
\packageIndices{NewPackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NewPackage}

Maintainer: \packageMaintainer{NewPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
